requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.7.0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
discord.py==1.7.3
dnspython==2.1.0
function==1.2.0
idna==3.2
multidict==5.1.0
pymongo==3.12.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
schedule==1.1.0
six==1.16.0
tzlocal==2.1
unittest-xml-reporting==3.0.4
urllib3==1.26.6
uwaterloodriver==1.1
yarl==1.6.3
